requirements.txt
# LiveKit Agent Dependencies
livekit>=0.11.0
livekit-agents>=0.8.0
livekit-plugins-openai>=0.6.0
livekit-plugins-silero>=0.6.0

# Utility
python-dotenv>=1.0.0

# Für HTTP-Requests (optional, für register-sip-trunk.py)
requests>=2.31.0

# JWT für API-Tokens (optional, für register-sip-trunk.py)
PyJWT>=2.8.0

